Fixed bug in anyElementInNS. Added tests.
Working through coverage tests. Fixed obscure bug in
anyElementInNS() that prevented it from selecting the base node of
a search. There is the possibility that this might need to be
rolled back, as the spec is not specific about whether the filter
ns|* should match the base node, or only its children. Currently
it matches the base node and descendants.