Fixed parser issues that were causing infinite loops.
Under certain circumstances, the tokenizer parser could keep
looping even though there was nothing to do. This happened when
legal but incorrect characters where positioned in certain places.
To fix this, I added a check in the main parsing loop to make sure
that progress is made on each parsing pass. Why I didn't do this
in the first place is beyond me.