Change in css function
Dear team of querypath: the css function needed a correction, so that it had the expected behavior when applying the css function. I compared this behaviour with JQuery and fortunately JQuery DOES behave in the expected matter, so querypath was the one behind. The error was as follows:
When making a css change, for example, we want to change a color:
 $qp($html, "p")->css("color", "blue");
but the behaviour was unbelievable: all of the "p" tags ended up with the same styles! If, for example, one <p> tag had a style of "margin-bottom: 2px", then this rule would spread in ALL of the other "<p>" tags even though we originally wanted to change only the color (!!). This was due to the way in which the loop and assignment of css style was assigned. 

Now the assignment is as it should be, and as it works in JQuery: ONLY CHANGE the specific attribute in the style of the css of each specific match element in the loop, without homogenizing the same style string for all of the matches.

Thanks, 

Best regards from Colombia, South America.

David López
http://investigacionyprogramacion.com