Fix composer.json schema
Fixes `$ composer validate`:
```
libraries/querypath/composer.json is valid for simple usage with composer but has
strict errors that make it unable to be published as a package:
See https://getcomposer.org/doc/04-schema.md for details on the schema
Name "querypath/QueryPath" does not match the best practice (e.g. lower-cased/with-dashes). We suggest using "querypath/query-path" instead. As such you will not be able to submit it to Packagist.
```